refactor: Line 생성자의 과도한 로직 분리(리뷰반영)
