+++ src/main/java/domain/Lines.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.line.Line;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.IntStream;
