+++ src/main/java/domain/line/LineValue.java
@@ -1,0 +1,9 @@
+package domain.line;
+
+import domain.BooleanGenerator;
+
+import java.util.List;
+
+public interface LineValue {
+    List<Point> value(BooleanGenerator generator);
+}
