+++ src/test/java/domain/PointTest.java
@@ -1,8 +1,8 @@
 package domain;
 
+import domain.line.Point;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
